Cleaning and Fixing
Cleaned up the lighting code and fixed a bug that I found. (line 126 in
Block.pde). When a lamp was under water but not under a solid block, it
would not light the blocks below it. This was because the code that was
controlling it was in an else if instead of an if.